Add mechanism for overriding with corresponding unittests.
